+++ packages/ui/useScroll/useScroll.js
@@ -119,7 +119,6 @@ const useScrollBase = (props) => {
 			horizontalScrollbar,
 			horizontalScrollbarHandle,
 			itemRenderer,
-			itemSize,
 			itemSizes,
 			noScrollByDrag,
 			noScrollByWheel,
@@ -151,6 +150,7 @@ const useScrollBase = (props) => {
 	delete rest.cbScrollTo;
 	delete rest.clearOverscrollEffect;
 	delete rest.handleResizeWindow;
+	delete rest.itemSize;
 	delete rest.onFlick;
 	delete rest.onKeyDown;
 	delete rest.onMouseDown;
@@ -173,6 +173,9 @@ const useScrollBase = (props) => {
 
 	const [isHorizontalScrollbarVisible, setIsHorizontalScrollbarVisible] = useState(horizontalScrollbar === 'visible');
 	const [isVerticalScrollbarVisible, setIsVerticalScrollbarVisible] = useState(verticalScrollbar === 'visible');
+	const [riRatio, setRiRatio] = useState(ri.scale(1));
+	const [originalItemSize, setOriginalItemSize] = useState(props.itemSize);
+	const [itemSize, setItemSize] = useState(props.itemSize);
 
 	const mutableRef = useRef({
 		overscrollEnabled: !!(props.applyOverscrollEffect),
@@ -297,6 +300,17 @@ const useScrollBase = (props) => {
 		}
 	});
 
+	if ((originalItemSize || originalItemSize === 0 || props.itemSize || props.itemSize === 0) && (originalItemSize !== props.itemSize)) {
+		setOriginalItemSize(props.itemSize);
+		if (typeof props.itemSize === 'object') {
+			if ((ri.scale(1) / riRatio !== props.itemSize.minWidth / itemSize.minWidth) || (ri.scale(1) / riRatio !== props.itemSize.minHeight / itemSize.minHeight)) {
+				setItemSize(props.itemSize);
+			}
+		} else if (ri.scale(1) / riRatio !== props.itemSize / itemSize) {
+			setItemSize(props.itemSize);
+		}
+	}
+
 	if (mutableRef.current.animator == null) {
 		mutableRef.current.animator = new ScrollAnimator();
 	}
@@ -397,8 +411,39 @@ const useScrollBase = (props) => {
 	}, [forceUpdate, scrollContentHandle]);
 	// scrollMode 'translate' ]]
 
+	const scrollContentProps = props.itemRenderer ? // If the child component is a VirtualList
+		{
+			childProps,
+			clientSize,
+			dataSize,
+			itemRenderer,
+			itemSize,
+			itemSizes,
+			overhang,
+			pageScroll,
+			spacing,
+			wrap
+		} :
+		{children};
+
 	function handleResizeWindow () {
 		const propsHandleResizeWindow = props.handleResizeWindow;
+
+		if (ri.scale(1) !== riRatio) {
+			if (scrollContentProps.itemSize.minWidth && scrollContentProps.itemSize.minHeight) {
+				scrollContentProps.itemSize.minWidth *= ri.scale(1) / riRatio;
+				scrollContentProps.itemSize.minHeight *= ri.scale(1) / riRatio;
+			} else {
+				scrollContentProps.itemSize *= ri.scale(1) / riRatio;
+				if (scrollContentProps.itemSizes) {
+					for (let i = 0; i < scrollContentProps.itemSizes.length; i++) {
+						scrollContentProps.itemSizes[i] *= ri.scale(1) / riRatio;
+					}
+				}
+			}
+			setItemSize(scrollContentProps.itemSize);
+			setRiRatio(ri.scale(1));
+		}
 
 		// `handleSize` in `ui/resolution.ResolutionDecorator` should be executed first.
 		setTimeout(() => {
@@ -411,7 +456,6 @@ const useScrollBase = (props) => {
 					scrollContentHandle.current.scrollToPosition(0, 0, 'instant');
 				}
 			}
-
 			enqueueForceUpdate();
 		});
 	}
@@ -842,6 +886,8 @@ const useScrollBase = (props) => {
 			}
 		} else {
 			props.preventScroll?.(ev);
+			mutableRef.current.repeat = ev.repeat;
+			if (ev.repeat && mutableRef.current.lastInputType === 'arrowKey') return;
 			forward('onKeyDown', ev, props);
 		}
 	}
@@ -1183,7 +1229,7 @@ const useScrollBase = (props) => {
 			let {roundedTargetX, roundedTargetY} = roundTarget(scrollContentHandle.current, targetX, targetY);
 
 			if (animate) {
-				scrollContentHandle.current.scrollToPosition(roundedTargetX, roundedTargetY, 'smooth');
+				scrollContentHandle.current.scrollToPosition(roundedTargetX, roundedTargetY, 'smooth', mutableRef.current.repeat);
 			} else {
 				scrollContentHandle.current.scrollToPosition(roundedTargetX, roundedTargetY, 'instant');
 			}
@@ -1576,21 +1622,6 @@ const useScrollBase = (props) => {
 			onTouchStart: scrollMode === 'native' ? onTouchStart : null // scrollMode 'native'
 		})
 	});
-
-	const scrollContentProps = props.itemRenderer ? // If the child component is a VirtualList
-		{
-			childProps,
-			clientSize,
-			dataSize,
-			itemRenderer,
-			itemSize,
-			itemSizes,
-			overhang,
-			pageScroll,
-			spacing,
-			wrap
-		} :
-		{children};
 
 	assignProperties('scrollContentProps', {
 		...scrollContentProps,