+++ CHANGELOG.md
@@ -2,11 +2,34 @@
 
 The following is a curated list of changes in the Enact project, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `core/util` function `checkPropTypes` to check the prop types of a component
+
+## [5.4.2] - 2026-01-28
+
+### Changed
+
+- `core/platform` to support `safari` 16.6, `chrome` 119, and `firefox` 128 or later
+- `ui/Scroller` scroll animation method for `scrollMode: native`
+
+### Fixed
+
+- `ui/VirtulList` not to cut off focused item when scroll by pressing key
+- `ui/VirtualList` not to jump to the top when item size was changed
+- `ui/VirtualList.VirtualList`, `ui/VirtualList.VirtualGridList` to adjust itemSize properly when resizing window
+
+## [5.4.1] - 2025-12-30
+
+No significant changes.
+
 ## [5.4.0] - 2025-12-08
 
 ### Added
 
-- `limestone/Card` `splitCaption` prop to allow user to split the content for the captions
+- `ui/Card` `splitCaption` prop to allow user to split the content for the captions
 - `ui/useScroll.ScrollAnimator` `oase-out-cubic` animation function
 
 ### Changed