+++ packages/sampler/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The following is a curated list of changes in the Enact Sampler, newest changes on the top.
 
+## [5.4.2] - 2026-01-28
+
+No significant changes.
+
+## [5.4.1] - 2025-12-30
+
+No significant changes.
+
 ## [5.4.0] - 2025-12-08
 
 No significant changes.